test(croquis): move setup-context scope regression test to its own module
